requirements.txt
altgraph==0.17
bidict==0.21.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
dnspython==1.16.0
docopt==0.6.2
eventlet==0.31.0
Flask==2.0.1
Flask-SocketIO==5.0.3
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
h11==0.12.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pyFirmata==1.1.0
PyQt5==5.15.0
PyQt5-sip==12.8.0
PyQt5-stubs==5.15.2.0
pyserial==3.4
python-engineio==4.2.0
python-socketio==5.3.0
requests==2.25.1
simple-websocket==0.2.0
six==1.16.0
urllib3==1.26.5
uvicorn==0.13.4
Werkzeug==2.0.1
wsproto==1.0.0
yarg==0.1.9
zope.event==4.5.0
zope.interface==5.4.0
